refactor: Entity와 JsonController에사용하는 Entity 클래스의 변수타입 변경 Integer -> Long

+++ src/main/java/com/example/demo/json/Entity.java
@@ -3,9 +3,9 @@ package com.example.demo.json;
 public class Entity {
 
     private String name;
-    private Integer age;
+    private Long age;
 
-    public Entity(Integer age, String name) {
+    public Entity(Long age, String name) {
         this.name = name;
         this.age = age;
     }
